mv jacobian graph and augmentation
